Rasberry Pi to Arduino
Because I could not post this from the Rpi, I uploaded everythig from home. Some of the changes made for this code wereto accomadate for differences in arduino versions i.e. windows to Linux. During Serial communication with Python, the Serial monitor may be innacurate and isn't a reliable method to determine what the code is doing so I added an LCD file that allows you to view variables. The Logging Data .ino was created with the intention of serial printing ALL data for troubleshooting purposes. The ReceivingFromPython code was the basis for serial communication in the master code. Serial Coms.py is the basic requirements for sending serial data from python to the arduino.